qc/modules/fastq_screen: consolidate pipeline tests for 'fastq_screen' QC module.
